refactor: move job fetching methods to own class
